refactor(upload planner): Rename types file PE-459
