+++ src/types/arfsdao_types.ts
@@ -17,7 +17,7 @@ import {
 	ArFSFileDataPrototype,
 	ArFSFileMetaDataPrototype
 } from '../exports';
-import { CreateDriveRewardSettings, UploadFileRewardSettings } from './cost_estimator_types';
+import { CreateDriveRewardSettings, UploadFileRewardSettings } from './upload_planner_types';
 import { TransactionID } from './transaction_id';
 
 /** Generic closure type that uses prepareArFSObjectTransaction (V2) or prepareArFSDataItem (bundle) */
